+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 3.0.0 (unreleased)
 ------------------
 
+* **2019-07-26**: [BC-BREAK] Remove deprecated `BaseTestCase::getClient`, use `BaseTestCase::getFrameworkBundleClient` instead.
 * **2019-02-28**: [BC-BREAK] Remove `DatabaseTestListener`, use scripts in `bin/make/` to ramp up your testing environment for several test suites
 * **2019-02-28**: [BC-BREAK] Introduce PHPUnit 6 Support, Remove PHPUnit 5.7.
 * **2019-02-28**: Add testing scenarios for Symfony 4.2 and 4.3 (allowed to fail for now)